+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGenerator.kt
@@ -2,6 +2,8 @@ package com.wutsi.codegen.kotlin.server
 
 import com.wutsi.codegen.Context
 import com.wutsi.codegen.core.generator.CodeGenerator
+import com.wutsi.codegen.editconfig.EditorConfigCodeGenerator
+import com.wutsi.codegen.github.GithubWorkflowCodeGenerator
 import com.wutsi.codegen.kotlin.KotlinMapper
 import com.wutsi.codegen.kotlin.sdk.SdkModelCodeGenerator
 import io.swagger.v3.oas.models.OpenAPI
@@ -13,5 +15,7 @@ class ServerCodeGenerator(private val mapper: KotlinMapper) : CodeGenerator {
         ServerControllerCodeGenerator(mapper).generate(openAPI, context)
         ServerPomCodeGenerator(mapper).generate(openAPI, context)
         ServerLauncherCodeGenerator().generate(openAPI, context)
+        EditorConfigCodeGenerator().generate(openAPI, context)
+        GithubWorkflowCodeGenerator().generate(openAPI, context)
     }
 }