+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGeneratorTest.kt
@@ -49,5 +49,12 @@ internal class ServerCodeGeneratorTest {
 
         // Pom
         assertTrue(File("${context.outputDirectory}/pom.xml").exists())
+
+        // .editorconfig
+        assertTrue(File("${context.outputDirectory}/.editorconfig").exists())
+
+        // Github Workflow
+        assertTrue(File("${context.outputDirectory}/.github/workflows/master.yml").exists())
+        assertTrue(File("${context.outputDirectory}/.github/workflows/pull_request.yml").exists())
     }
 }